Move platform unzip test
