+++ util/unzip_test.go
@@ -6,11 +6,8 @@ package util
 import (
 	"fmt"
 	"os"
-	"os/user"
 	"path/filepath"
 	"runtime"
-	"strconv"
-	"syscall"
 	"testing"
 	"time"
 
@@ -151,29 +148,6 @@ func TestUnzipOverMoveExisting(t *testing.T) {
 	assert.NoError(t, err)
 
 	assertFileExists(t, filepath.Join(tmpDir, filepath.Base(destinationPath)))
-}
-
-// TestUnzipOtherUser checks to make sure that a zip file created from a
-// different uid has the current uid after unpacking.
-func TestUnzipOtherUser(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("Unsupported on windows")
-	}
-	var testZipOtherUserPath = filepath.Join(os.Getenv("GOPATH"), "src/github.com/keybase/go-updater/test/test-uid-503.zip")
-	destinationPath := TempPath("", "TestUnzipOtherUser.")
-	err := Unzip(testZipOtherUserPath, destinationPath, testLog)
-	require.NoError(t, err)
-
-	// Get uid, gid of current user
-	currentUser, err := user.Current()
-	require.NoError(t, err)
-	uid, err := strconv.Atoi(currentUser.Uid)
-	require.NoError(t, err)
-
-	fileInfo, err := os.Stat(filepath.Join(destinationPath, "test"))
-	require.NoError(t, err)
-	fileUID := fileInfo.Sys().(*syscall.Stat_t).Uid
-	assert.Equal(t, uid, int(fileUID))
 }
 
 // TestUnzipFileModTime checks to make sure after unpacking zip file the file
